Project cleaned : include and src folders added to editor project, classes added to gameStartup project
