+++ editor/include/Editor/Editor.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+class Editor
+{
+public:
+    void update()
+    {
+
+    }
+};
